[SELC-8576] feat: create paty detail page
