+++ src/pages/adminPage/AdminPage.tsx
@@ -1,47 +1,16 @@
-import { ArrowForward } from '@mui/icons-material';
 import SearchIcon from '@mui/icons-material/Search';
-import {
-  Autocomplete,
-  Chip,
-  CircularProgress,
-  Divider,
-  Grid,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  TextField,
-  Typography,
-} from '@mui/material';
-import { ButtonNaked, PartyAccountItemButton } from '@pagopa/mui-italia';
-import { TitleBox, useErrorDispatcher, usePermissions } from '@pagopa/selfcare-common-frontend/lib';
-import { setProductPermissions } from '@pagopa/selfcare-common-frontend/lib/redux/slices/permissionsSlice';
+import { Autocomplete, CircularProgress, Grid, TextField, Typography } from '@mui/material';
+import { PartyAccountItemButton } from '@pagopa/mui-italia';
+import { TitleBox, useErrorDispatcher } from '@pagopa/selfcare-common-frontend/lib';
 import { trackEvent } from '@pagopa/selfcare-common-frontend/lib/services/analyticsService';
-import { Actions } from '@pagopa/selfcare-common-frontend/lib/utils/constants';
 import { storageOpsBuilder } from '@pagopa/selfcare-common-frontend/lib/utils/storage-utils';
 import { debounce } from 'lodash';
-import { Fragment, useEffect, useMemo, useState } from 'react';
-import { Trans, useTranslation } from 'react-i18next';
-import { ProductOnBoardingStatusEnum } from '../../api/generated/b4f-dashboard/OnboardedProductResource';
+import { useEffect, useMemo, useState } from 'react';
+import { useTranslation } from 'react-i18next';
 import { SearchServiceInstitution } from '../../api/generated/party-registry-proxy/SearchServiceInstitution';
-import { Party } from '../../model/Party';
-import { Product } from '../../model/Product';
-import { useAppDispatch } from '../../redux/hooks';
-import { fetchPartyDetailsService } from '../../services/dashboardService';
 import { searchInstitutionsService } from '../../services/partyRegistryProxyService';
-import { fetchProducts } from '../../services/productService';
 import { buildUrlLog } from '../../utils/helper';
-import AdminPartyInfo from './components/AdminPartyInfo';
-import BackofficeNotIntegratedModal from './components/BackofficeNotIntegratedModal';
-import GenericEnvProductModal from './components/GenericEnvProductModal';
-import ProductAvatarCell from './components/ProductAvatarCell';
-import SessionModalInteropProduct from './components/SessionModalInteropProduct';
-import { useProductFiltering } from './hooks/useProductFiltering';
-import { useProductNavigation } from './hooks/useProductNavigation';
 import { commonStyles, CustomListbox } from './utils/styles';
-import { isProductAllowed } from './utils/utils';
 
 const AdminPage = () => {
   const [loading, setLoading] = useState(false);
@@ -50,53 +19,12 @@ const AdminPage = () => {
   const [selectedInstitution, setSelectedInstitution] = useState<SearchServiceInstitution | null>(
     null
   );
-  const [partyDetail, setPartyDetail] = useState<Party | null>(null);
-  const [products, setProducts] = useState<Array<Product>>([]);
-  const [openBackofficeNotIntegratedModal, setOpenBackofficeNotIntegratedModal] = useState(false);
 
   const { t } = useTranslation();
   const addError = useErrorDispatcher();
-  const dispatch = useAppDispatch();
-  const { hasPermission } = usePermissions();
-
-  const {
-    productsToShow,
-    interopProductsList,
-    hasMoreThanOneInteropEnv,
-    getProductTitle,
-    getActiveSubProduct,
-  } = useProductFiltering({ partyDetail, products });
-
-  const {
-    activeProduct,
-    interopProduction,
-    openInteropModal,
-    openGenericEnvModal,
-    handleOnboardedProductClick,
-    handleInteropConfirm,
-    handleGenericEnvConfirm,
-    closeInteropModal,
-    closeGenericEnvModal,
-  } = useProductNavigation({ products, selectedInstitution, hasMoreThanOneInteropEnv });
 
   useEffect(() => {
     trackEvent('BACKSTAGE_DASHBOARD');
-  }, []);
-
-  useEffect(() => {
-    fetchProducts()
-      .then((products) => {
-        setProducts(products);
-      })
-      .catch((error) => {
-        addError({
-          id: 'fetchProducts-api-error',
-          blocking: false,
-          techDescription: 'Fetch products failed',
-          toNotify: false,
-          error: error as Error,
-        });
-      });
   }, []);
 
   const debouncedSearch = useMemo(
@@ -149,36 +77,6 @@ const AdminPage = () => {
       setSelectedInstitution(storedInstitution);
     }
   }, []);
-
-  useEffect(() => {
-    if (selectedInstitution?.id) {
-      fetchPartyDetailsService(selectedInstitution.id)
-        .then((party) => {
-          if (party) {
-            setPartyDetail(party);
-            const productPermissions = [...party.products]
-              .filter((product) => product.productOnBoardingStatus === 'ACTIVE')
-              .map((product) => ({
-                productId: product.productId ?? '',
-                actions: product.userProductActions ? [...product.userProductActions] : [],
-              }));
-            dispatch(setProductPermissions(productPermissions));
-          }
-        })
-        .catch((error) => {
-          storageOpsBuilder('selectedInstitution', 'object', false).delete();
-          setSelectedInstitution(null);
-
-          addError({
-            id: `fetchPartyDetails-${selectedInstitution.id}-api-error`,
-            blocking: false,
-            techDescription: `Fetch party details for institution id: ${selectedInstitution.id} failed`,
-            toNotify: false,
-            error: error as Error,
-          });
-        });
-    }
-  }, [selectedInstitution]);
 
   return (
     <Grid px={3} mt={3} sx={{ width: '100%' }}>
@@ -286,156 +184,6 @@ const AdminPage = () => {
           }}
         />
       </Grid>
-
-      {partyDetail && selectedInstitution && (
-        <Grid item xs={12} sx={commonStyles}>
-          <AdminPartyInfo partyDetail={partyDetail} selectedInstitution={selectedInstitution} />
-          <Divider sx={{ my: 3 }} />
-
-          {productsToShow && productsToShow.length > 0 && (
-            <>
-              <TableContainer>
-                <Table>
-                  <TableHead>
-                    <TableRow>
-                      <TableCell sx={{ fontWeight: 600 }}>
-                        {t('adminPage.selectedPartyDetails.product')}
-                      </TableCell>
-                      <TableCell sx={{ fontWeight: 600 }}>
-                        {t('adminPage.selectedPartyDetails.subscriptionDate')}
-                      </TableCell>
-                      <TableCell sx={{ fontWeight: 600 }}>
-                        {t('adminPage.selectedPartyDetails.agreementStatus')}
-                      </TableCell>
-                      <TableCell sx={{ fontWeight: 600 }}>
-                        {t('adminPage.selectedPartyDetails.institutionType')}
-                      </TableCell>
-                      <TableCell sx={{ fontWeight: 600 }} align="right"></TableCell>
-                    </TableRow>
-                  </TableHead>
-                  <TableBody>
-                    {productsToShow?.map((onboardedProduct) => {
-                      const productFromConfiguration = products.find(
-                        (p) => p.id === onboardedProduct?.productId
-                      );
-
-                      if (!productFromConfiguration) {
-                        return null;
-                      }
-
-                      return (
-                        <Fragment key={onboardedProduct?.productId}>
-                          <TableRow hover>
-                            <TableCell>
-                              <ProductAvatarCell
-                                onboardedProduct={onboardedProduct}
-                                productFromConfiguration={productFromConfiguration}
-                                getActiveSubProduct={getActiveSubProduct}
-                                getProductTitle={getProductTitle}
-                              />
-                            </TableCell>
-                            <TableCell>
-                              {onboardedProduct?.createdAt
-                                ? new Date(onboardedProduct.createdAt).toLocaleDateString()
-                                : '-'}
-                            </TableCell>
-                            <TableCell>
-                              <Chip
-                                label={t('adminPage.selectedPartyDetails.activeStatus')}
-                                size="small"
-                                color="success"
-                                sx={{ backgroundColor: 'success.light', color: 'success.main' }}
-                              />
-                            </TableCell>
-                            <TableCell>
-                              {t(
-                                `common.institutionType.descriptions.${onboardedProduct?.institutionType?.toLowerCase()}`
-                              ) || '-'}
-                            </TableCell>
-                            {onboardedProduct.productOnBoardingStatus ===
-                              ProductOnBoardingStatusEnum.ACTIVE &&
-                              hasPermission(
-                                onboardedProduct.productId || '',
-                                Actions.AccessProductBackofficeAdmin
-                              ) && (
-                                <TableCell align="right">
-                                  <ButtonNaked
-                                    component="button"
-                                    endIcon={<ArrowForward />}
-                                    onClick={() => {
-                                      trackEvent('BACKSTAGE_BACK_OFFICE_CLICK', {
-                                        product_id: onboardedProduct.productId || '',
-                                      });
-                                      if (isProductAllowed(onboardedProduct.productId || '')) {
-                                        handleOnboardedProductClick(productFromConfiguration);
-                                      } else {
-                                        setOpenBackofficeNotIntegratedModal(true);
-                                      }
-                                    }}
-                                    sx={{ color: 'primary.main', fontWeight: 'bold' }}
-                                  >
-                                    {t('adminPage.selectedPartyDetails.backOffice')}
-                                  </ButtonNaked>
-                                </TableCell>
-                              )}
-                          </TableRow>
-                        </Fragment>
-                      );
-                    })}
-                  </TableBody>
-                </Table>
-              </TableContainer>
-              <BackofficeNotIntegratedModal
-                open={openBackofficeNotIntegratedModal}
-                productName={activeProduct?.title ?? ''}
-                onClose={() => setOpenBackofficeNotIntegratedModal(false)}
-              />
-              <SessionModalInteropProduct
-                open={openInteropModal}
-                title={t('overview.activeProducts.activeProductsEnvModal.title')}
-                message={
-                  <Trans
-                    i18nKey="overview.activeProducts.activeProductsEnvModal.message"
-                    values={{
-                      productTitle: activeProduct?.id?.startsWith('prod-interop')
-                        ? interopProduction?.title
-                        : activeProduct?.title,
-                    }}
-                    components={{ 1: <strong /> }}
-                  >
-                    {`Sei stato abilitato ad operare negli ambienti riportati di seguito per il prodotto <1>{{productTitle}}</1>.`}
-                  </Trans>
-                }
-                onConfirmLabel={t('overview.activeProducts.activeProductsEnvModal.enterButton')}
-                onCloseLabel={t('overview.activeProducts.activeProductsEnvModal.backButton')}
-                onConfirm={handleInteropConfirm}
-                handleClose={closeInteropModal}
-                authorizedInteropProducts={interopProductsList?.map((p) => p.productId || '')}
-                products={products}
-                party={selectedInstitution}
-              />
-              <GenericEnvProductModal
-                open={openGenericEnvModal}
-                title={t('overview.activeProducts.activeProductsEnvModal.title')}
-                message={
-                  <Trans
-                    i18nKey="overview.activeProducts.activeProductsEnvModal.message"
-                    values={{ productTitle: activeProduct?.title }}
-                    components={{ 1: <strong /> }}
-                  >
-                    {`Sei stato abilitato ad operare negli ambienti riportati di seguito per il prodotto <1>{{productTitle}}</1>.`}
-                  </Trans>
-                }
-                onConfirmLabel={t('overview.activeProducts.activeProductsEnvModal.enterButton')}
-                onCloseLabel={t('overview.activeProducts.activeProductsEnvModal.backButton')}
-                onConfirm={(e) => handleGenericEnvConfirm((e.target as HTMLInputElement).value)}
-                handleClose={closeGenericEnvModal}
-                productEnvironments={activeProduct?.backOfficeEnvironmentConfigurations as any}
-              />
-            </>
-          )}
-        </Grid>
-      )}
     </Grid>
   );
 };
