+++ src/pages/adminPage/hooks/useProductNavigation.ts
@@ -1,18 +1,18 @@
 import i18n from '@pagopa/selfcare-common-frontend/lib/locale/locale-utils';
 import { useState } from 'react';
-import { SearchServiceInstitution } from '../../../api/generated/party-registry-proxy/SearchServiceInstitution';
 import { useTokenExchange } from '../../../hooks/useTokenExchange';
+import { Party } from '../../../model/Party';
 import { Product } from '../../../model/Product';
 
 type UseProductNavigationProps = {
   products: Array<Product>;
-  selectedInstitution: SearchServiceInstitution | null;
+  partyDetail: Party | null;
   hasMoreThanOneInteropEnv: boolean;
 };
 
 export const useProductNavigation = ({
   products,
-  selectedInstitution,
+  partyDetail,
   hasMoreThanOneInteropEnv,
 }: UseProductNavigationProps) => {
   const [activeProduct, setActiveProduct] = useState<Product | null>(null);
@@ -45,30 +45,15 @@ export const useProductNavigation = ({
       return;
     }
 
-    void invokeProductBo(
-      productFromConfiguration,
-      selectedInstitution as SearchServiceInstitution,
-      undefined,
-      lang
-    );
+    void invokeProductBo(productFromConfiguration, partyDetail, undefined, lang);
   };
 
   const handleInteropConfirm = () => {
-    void invokeProductBo(
-      interopProduction as Product,
-      selectedInstitution as SearchServiceInstitution,
-      undefined,
-      lang
-    );
+    void invokeProductBo(interopProduction as Product, partyDetail, undefined, lang);
   };
 
   const handleGenericEnvConfirm = (envValue: string) => {
-    void invokeProductBo(
-      activeProduct as Product,
-      selectedInstitution as SearchServiceInstitution,
-      envValue,
-      lang
-    );
+    void invokeProductBo(activeProduct as Product, partyDetail, envValue, lang);
   };
 
   return {
