+++ src/utils/constants.ts
@@ -6,3 +6,4 @@ export const LOG_REDUX_ACTIONS = IS_DEVELOP;
 export const LOADING_RETRIEVE_ONBOARDING_REQUEST = 'RETRIEVE_ONBOARDING_REQUEST';
 export const LOADING_RETRIEVE_INSTITUTIONS = 'RETRIEVE_INSTITUTIONS';
 export const LOADING_TASK_TOKEN_EXCHANGE = 'TOKEN_EXCHANGE';
+export const LOADING_TASK_PRODUCTS = 'LOADING_TASK_PRODUCTS';
